# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API WorkLink",
        "description": "Documentación de la API REST",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refrescar token",
                "description": "Genera un nuevo token JWT",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "Token refrescado exitosamente"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/availabilities": {
            "get": {
                "tags": [
                    "Availabilities"
                ],
                "summary": "Listar disponibilidades",
                "description": "Retorna todas las disponibilidades con perfil freelancer, usuario y rol.",
                "operationId": "listAvailabilities",
                "responses": {
                    "200": {
                        "description": "Disponibilidades obtenidas exitosamente"
                    },
                    "401": {
                        "description": "No autorizado. Token requerido o inválido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Availabilities"
                ],
                "summary": "Crear disponibilidad",
                "description": "Crea un rango de disponibilidad para un perfil freelancer. Un freelancer crea disponibilidad para su propio perfil; un admin puede indicar freelancer_id.",
                "operationId": "createAvailability",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "freelancer_id": {
                                        "description": "Solo requerido si el usuario autenticado es admin",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-15"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-30"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "busy",
                                            "vacation"
                                        ],
                                        "example": "available"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Disponibilidad creada exitosamente"
                    },
                    "401": {
                        "description": "No autorizado. Token requerido o inválido"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/availabilities/{id}": {
            "get": {
                "tags": [
                    "Availabilities"
                ],
                "summary": "Obtener disponibilidad por ID",
                "description": "Retorna los detalles de una disponibilidad.",
                "operationId": "showAvailability",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la disponibilidad",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Disponibilidad obtenida exitosamente"
                    },
                    "401": {
                        "description": "No autorizado. Token requerido o inválido"
                    },
                    "404": {
                        "description": "Disponibilidad no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Availabilities"
                ],
                "summary": "Actualizar disponibilidad",
                "description": "Actualiza las fechas o el estado de una disponibilidad existente.",
                "operationId": "updateAvailability",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la disponibilidad",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-15"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-10"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "busy",
                                            "vacation"
                                        ],
                                        "example": "busy"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Disponibilidad actualizada correctamente"
                    },
                    "401": {
                        "description": "No autorizado. Token requerido o inválido"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "Disponibilidad no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Availabilities"
                ],
                "summary": "Eliminar disponibilidad",
                "description": "Elimina una disponibilidad existente.",
                "operationId": "deleteAvailability",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la disponibilidad",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Disponibilidad eliminada correctamente"
                    },
                    "401": {
                        "description": "No autorizado. Token requerido o inválido"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "Disponibilidad no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}